Fix tests/lint
