+++ packages/json-api/src/-private/cache.ts
@@ -4,7 +4,7 @@
 import type { CollectionEdge, Graph, GraphEdge, ImplicitEdge, ResourceEdge } from '@ember-data/graph/-private';
 import { graphFor, isBelongsTo, peekGraph } from '@ember-data/graph/-private';
 import type Store from '@ember-data/store';
-import type { CacheCapabilitiesManager, SchemaService } from '@ember-data/store/types';
+import type { CacheCapabilitiesManager } from '@ember-data/store/types';
 import { LOG_MUTATIONS, LOG_OPERATIONS, LOG_REQUESTS } from '@warp-drive/build-config/debugging';
 import { DEPRECATE_RELATIONSHIP_REMOTE_UPDATE_CLEARING_LOCAL_STATE } from '@warp-drive/build-config/deprecations';
 import { DEBUG } from '@warp-drive/build-config/env';
@@ -48,6 +48,8 @@ import type {
   SingleResourceDocument,
   SingleResourceRelationship,
 } from '@warp-drive/core-types/spec/json-api-raw';
+
+import { validateDocumentFields } from './validate-document-fields';
 
 type IdentifierCache = Store['identifierCache'];
 type InternalCapabilitiesManager = CacheCapabilitiesManager & { _store: Store };
@@ -1970,109 +1972,5 @@ function copyLinksAndMeta(target: { links?: unknown; meta?: unknown }, source: o
   }
   if ('meta' in source) {
     target.meta = source.meta;
-  }
-}
-
-function validateDocumentFields(
-  schema: SchemaService,
-  jsonApiDoc: SingleResourceDocument | CollectionResourceDocument
-) {
-  const { data, included } = jsonApiDoc;
-  if (data === null) {
-    return;
-  }
-
-  if (Array.isArray(data)) {
-    for (const resource of data) {
-      validateResourceFields(schema, resource, { verifyIncluded: true, included });
-    }
-  } else {
-    validateResourceFields(schema, data, { verifyIncluded: true, included });
-  }
-
-  if (included) {
-    for (const resource of included) {
-      validateResourceFields(schema, resource, { verifyIncluded: false });
-    }
-  }
-}
-
-function validateResourceFields(
-  schema: SchemaService,
-  resource: ExistingResourceObject,
-  options:
-    | {
-        verifyIncluded: true;
-        included: ExistingResourceObject[] | undefined;
-      }
-    | {
-        verifyIncluded: false;
-      }
-) {
-  if (!resource.relationships) {
-    return;
-  }
-
-  const resourceType = resource.type;
-  const fields = schema.fields({ type: resource.type });
-  for (const [type, relationshipDoc] of Object.entries(resource.relationships)) {
-    const field = fields.get(type);
-    if (!field) {
-      return;
-    }
-    switch (field.kind) {
-      case 'belongsTo': {
-        if (field.options.async) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the field is in linksMode but async is not yet supported`
-          );
-        }
-
-        if (!relationshipDoc.links?.related) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the field is in linksMode but the related link is missing`
-          );
-        }
-
-        const relationshipData = relationshipDoc.data;
-        if (Array.isArray(relationshipData)) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the relationship data for a belongsTo relationship is unexpectedly an array`
-          );
-        }
-        // Explicitly allow `null`! Missing key or `undefined` are always invalid.
-        if (relationshipData === undefined) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the field is in linksMode but the relationship data is undefined`
-          );
-        }
-        if (relationshipData === null) {
-          return;
-        }
-
-        if (!options.verifyIncluded) {
-          return;
-        }
-        const includedDoc = options.included?.find(
-          (doc) => doc.type === relationshipData.type && doc.id === relationshipData.id
-        );
-        if (!includedDoc) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the field is in linksMode but the related data is not included`
-          );
-        }
-        break;
-      }
-      case 'hasMany': {
-        if (field.options.linksMode) {
-          throw new Error(
-            `Cannot fetch ${resourceType}.${field.name} because the field is in linksMode but hasMany is not yet supported`
-          );
-        }
-        break;
-      }
-      default:
-        break;
-    }
   }
 }
