+++ tests/warp-drive__schema-record/tests/legacy/reactivity/relationships-test.ts
@@ -900,7 +900,7 @@ module('Legacy | Reactivity | relationships', function (hooks) {
     const Wes = store.peekRecord<User>('user', '3')!;
     const resource = schema.resource({ type: 'user' });
 
-    const { counters, fieldOrder } = await reactiveContext.call(this, Rey, resource, {
+    const { counters, fieldOrder } = await reactiveContext(Rey, resource, {
       bestFriend: 'name',
     });
