+++ tests/warp-drive__schema-record/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "include": ["app/**/*", "config/**/*", "tests/**/*"],
+  "include": ["app/**/*", "config/**/*", "tests/**/*", "../../@types/ember-data-qunit-asserts"],
   "glint": {
     "environment": ["ember-loose", "ember-template-imports"]
   },
